+++ AST/Variable_Storage.java
@@ -7,7 +7,7 @@ import java.util.HashMap;
 public class Variable_Storage {
 
     private static Variable_Storage instance;
-    private final HashMap<String,Double>[] PlayerVariable ;
+    private final HashMap<String,Long>[] PlayerVariable ;
     private Command findPlayer = Command.instance();
 
     private Variable_Storage(){
@@ -15,6 +15,7 @@ public class Variable_Storage {
         for(int i = 0; i < Command.instance().game.ListOfPlayer.length ; i++)
         this.PlayerVariable[i] = new HashMap<>();
     }
+
     public static Variable_Storage instance(){
         if(instance == null){
             instance = new Variable_Storage();
@@ -26,14 +27,14 @@ public class Variable_Storage {
         int curr_player = findPlayer.game.cur_player;
         PlayerVariable[curr_player].put(_VarName,_VarValue.eval(PlayerVariable[curr_player]));
     }
-    public void AssignVariable(String _VarName,double _VarValue){
+    public void AssignVariable(String _VarName,Long _VarValue){
         int curr_player = findPlayer.game.cur_player;
         PlayerVariable[curr_player].put(_VarName,_VarValue);
     }
     public double GetVariableValue(String key){
         return PlayerVariable[findPlayer.game.cur_player].get(key);
     }
-    public HashMap<String,Double> GetVariableMap(){
+    public HashMap<String,Long> GetVariableMap(){
         return PlayerVariable[findPlayer.game.cur_player];
     }
 }
